+++ Frends.LinkedIn.SearchAdAccountUser/CHANGELOG.md
@@ -3,3 +3,7 @@
 ## [1.0.0] - 2023-10-23
 ### Added
 - Initial implementation
+
+## [1.0.1] - 2023-10-31
+### Added
+- Added multitagreting to net471 and netstandard2.0
