+++ Frends.LinkedIn.SearchAdAccountUser/Frends.LinkedIn.SearchAdAccountUser/GlobalSuppressions.cs
@@ -10,3 +10,4 @@
 [assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1602:Enumeration items should be documented", Justification = "Enumerations need no docs", Scope = "namespaceanddescendants", Target = "~N:Frends.LinkedIn.SearchAdAccountUser.Definitions")]
 [assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1629:Documentation text should end with a period", Justification = "Following Frends documentation guidelines", Scope = "namespaceanddescendants", Target = "~N:Frends.LinkedIn.SearchAdAccountUser")]
 [assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1633:File should have header", Justification = "Following Frends documentation guidelines")]
+[assembly: SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1649:File name should match first type name", Justification = "Following Frends documentation guidelines", Scope = "namespaceanddescendants", Target = "~N:Frends.LinkedIn.SearchAdAccountUser")]
